+++ plugins/SimpleAR/src/Main.cpp
@@ -236,7 +236,7 @@ INT addEvent(WPARAM hContact, LPARAM hDBEvent)
 						dbei.flags = DBEF_UTF | DBEF_SENT; //DBEF_READ;
 						dbei.szModule = pszProto;
 						dbei.timestamp = time(NULL);
-						dbei.cbBlob = mir_strlen(pszUtf) + 1;
+						dbei.cbBlob = (int)mir_strlen(pszUtf) + 1;
 						dbei.pBlob = (PBYTE)pszUtf;
 						db_event_add(hContact, &dbei);
